backend/apps/common/urls.py
from django.urls import path
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny
from rest_framework.response import Response


@api_view(['GET'])
@permission_classes([AllowAny])
def health_check(request):
    """Endpoint de health check - público, não requer autenticação"""
    return Response({'status': 'ok'})


urlpatterns = [
    path('health/', health_check, name='health'),
]

